# solution/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AdsPlatform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/clients/{client_id}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Get Client",
        "operationId": "get_client_clients__client_id__get",
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clients/bulk": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Bulk Update",
        "operationId": "bulk_update_clients_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/Client"
                },
                "type": "array",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  },
                  "type": "array",
                  "title": "Response Bulk Update Clients Bulk Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/advertisers/{advertiser_id}": {
      "get": {
        "tags": [
          "Advertisers"
        ],
        "summary": "Get Advertiser",
        "operationId": "get_advertiser_advertisers__advertiser_id__get",
        "parameters": [
          {
            "name": "advertiser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Advertiser Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Advertiser"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/advertisers/{advertiser_id}/campaigns": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create Campaign",
        "operationId": "create_campaign_advertisers__advertiser_id__campaigns_post",
        "parameters": [
          {
            "name": "advertiser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Advertiser Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get Campaigns",
        "operationId": "get_campaigns_advertisers__advertiser_id__campaigns_get",
        "parameters": [
          {
            "name": "advertiser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Advertiser Id"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Size"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "default": 0,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Campaign"
                  },
                  "title": "Response Get Campaigns Advertisers  Advertiser Id  Campaigns Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/advertisers/{advertiser_id}/campaigns/{campaign_id}": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get Campaign",
        "operationId": "get_campaign_advertisers__advertiser_id__campaigns__campaign_id__get",
        "parameters": [
          {
            "name": "advertiser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Advertiser Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Update Campaign",
        "operationId": "update_campaign_advertisers__advertiser_id__campaigns__campaign_id__put",
        "parameters": [
          {
            "name": "advertiser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Advertiser Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Delete Campaign",
        "operationId": "delete_campaign_advertisers__advertiser_id__campaigns__campaign_id__delete",
        "parameters": [
          {
            "name": "advertiser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Advertiser Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Advertiser": {
        "properties": {
          "advertiser_id": {
            "type": "string",
            "format": "uuid",
            "title": "Advertiser Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "advertiser_id",
          "name"
        ],
        "title": "Advertiser"
      },
...
      "Campaign": {
        "properties": {
          "impressions_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Impressions Limit"
          },
          "clicks_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Clicks Limit"
          },
          "cost_per_impression": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cost Per Impression"
          },
          "cost_per_click": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cost Per Click"
          },
          "start_date": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Date"
          },
          "end_date": {
            "type": "integer",
            "minimum": 0.0,
            "title": "End Date"
          },
          "ad_title": {
            "type": "string",
            "title": "Ad Title"
          },
          "ad_text": {
            "type": "string",
            "title": "Ad Text"
          },
          "ad_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Image"
          },
          "targeting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CampaignTargeting"
              },
              {
                "type": "null"
              }
            ]
          },
          "campaign_id": {
            "type": "string",
            "format": "uuid",
            "title": "Campaign Id"
          },
          "advertiser_id": {
            "type": "string",
            "format": "uuid",
            "title": "Advertiser Id"
          }
        },
        "type": "object",
        "required": [
          "impressions_limit",
          "clicks_limit",
          "cost_per_impression",
          "cost_per_click",
          "start_date",
          "end_date",
          "ad_title",
          "ad_text",
          "campaign_id",
          "advertiser_id"
        ],
        "title": "Campaign"
      },
...
      "CampaignEdit": {
        "properties": {
          "impressions_limit": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Impressions Limit"
          },
          "clicks_limit": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Clicks Limit"
          },
          "cost_per_impression": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Per Impression"
          },
          "cost_per_click": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Per Click"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "ad_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Title"
          },
          "ad_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Text"
          },
          "ad_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Image"
          },
          "targeting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CampaignTargeting"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "CampaignEdit"
      },
      "CampaignIn": {
        "properties": {
          "impressions_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Impressions Limit"
          },
          "clicks_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Clicks Limit"
          },
          "cost_per_impression": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cost Per Impression"
          },
          "cost_per_click": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cost Per Click"
          },
          "start_date": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Date"
          },
          "end_date": {
            "type": "integer",
            "minimum": 0.0,
            "title": "End Date"
          },
          "ad_title": {
            "type": "string",
            "title": "Ad Title"
          },
          "ad_text": {
            "type": "string",
            "title": "Ad Text"
          },
          "ad_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Image"
          },
          "targeting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CampaignTargeting"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "impressions_limit",
          "clicks_limit",
          "cost_per_impression",
          "cost_per_click",
          "start_date",
          "end_date",
          "ad_title",
          "ad_text"
        ],
        "title": "CampaignIn"
      },
...
      "CampaignTargeting": {
        "properties": {
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetingGender"
              },
              {
                "type": "null"
              }
            ]
          },
          "age_from": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age From"
          },
          "age_to": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age To"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          }
        },
        "type": "object",
        "title": "CampaignTargeting"
      },
      "Client": {
        "properties": {
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          },
          "login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Login"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "client_id"
        ],
        "title": "Client"
      },
      "Gender": {
        "type": "string",
        "enum": [
          "MALE",
          "FEMALE"
        ],
        "title": "Gender"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TargetingGender": {
        "type": "string",
        "enum": [
          "MALE",
          "FEMALE",
          "ALL"
        ],
        "title": "TargetingGender"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}